Inicio Stock quotes service

+++ Cotacao/Cotacao.Adapter/Extensions/JsonExtensions.cs
@@ -13,10 +13,10 @@ namespace Cotacao.Adapter.Extensions
             var entry = new TimeSeries
             {
                 Timestamp = date,
-                Open = token.First.Value<double>("1. open"),
-                High = token.First.Value<double>("2. high"),
-                Low = token.First.Value<double>("3. low"),
-                Close = token.First.Value<double>("4. close"),
+                Open = token.First.Value<float>("1. open"),
+                High = token.First.Value<float>("2. high"),
+                Low = token.First.Value<float>("3. low"),
+                Close = token.First.Value<float>("4. close"),
                 Volume = token.First.Value<long>("5. volume")
             };
