+++ Cotacao/Cotacao.Adapter/Models/TimeSeries.cs
@@ -5,10 +5,10 @@ namespace Cotacao.Adapter.Models
     public class TimeSeries
     {
         public DateTime Timestamp { get; set; }
-        public double Open { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Close { get; set; }
+        public float Open { get; set; }
+        public float High { get; set; }
+        public float Low { get; set; }
+        public float Close { get; set; }
         public long Volume { get; set; }
     }
 }
